Fix cast warning

+++ src/user/common.cpp
@@ -10,7 +10,7 @@ char **kiv_os::getArgs(const kiv_os::TRegisters &context, int *argc) {
 	kiv_os::TProcess_Startup_Info *info = (kiv_os::TProcess_Startup_Info *)context.rdi.r;
 	std::string line = info->arg;
 	std::vector<std::string> parts = parseLine(line);
-	*argc = parts.size();
+	*argc = static_cast<int>(parts.size());
 	std::vector<char *> argv{};
 	for (size_t i = 0; i < *argc; i++) {
 		argv.push_back(&parts[i].front());
